fix: Fix dispatch event checkout (#3407)
<!-- Feel free to delete comments as you fill this in -->
- Fix checking out code in the repository dispatch event by specifying
the repository and ref.
- Print more debug info during hash verification.
<!-- summary of changes -->

## Test Plan
Dispatch events utilize pipelines from the default branch, and custom
dispatch events cannot be run before merge—see
https://github.com/orgs/community/discussions/24657?sort=top. This is
why this change was not tested in this repository, as the changes
already need to be present on the default branch.

Instead, I tested it on my personal repository with a simplified setup,
but preserving the logic. I confirmed that for dispatch events, the
checkout was done to the default origin branch, and after these changes,
the checkout is done on the correct sha in the forked branch.

## References

https://github.com/Snowflake-Labs/terraform-provider-snowflake/pull/3400#issuecomment-2666395813
## TODO
Check the dispatch after it's merged to the default branch.