chore: Handle read operation safely (#3491)
## Changes
- Added a common read function that is safely calling show by id, and in
case of any issues with querying the object, it checks higher hierarchy
objects to see if the object is connected to them being missing or not.
- New function is intended to be used inside newly introduced
ShowByIDSafely functions on the SDK level.
- I have added a new common Snowflake error because we can use it with a
combination of other errors to check whether the underlying database or
schema objects are missing.
- I left the client changes (I was using them in the previous approach)
because they're following the conventions of other testing clients.

## Next changes
- Apply new show by id function to all resources / data sources.
- Adjust SDK generator templates to include new function (or add readme
to do it later)
- Check object renaming tests and adjust with current Read handling