chore: Run tests only for changes in certain files (#3550)
`paths-ignore` does not work well with workflows that are required to
pass
([docs](https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#example-including-paths),
[example thread](https://github.com/orgs/community/discussions/54877),
[our
tests](https://github.com/snowflakedb/terraform-provider-snowflake/pull/3542)).

Because of that, we are switching to `dorny/paths-filter@v3`
(https://github.com/marketplace/actions/paths-changes-filter).

With the current set up we require changes in one of the listed
directiories/files to trigger the tests. Experiment PRs showing the
behavior are available here:
- https://github.com/snowflakedb/terraform-provider-snowflake/pull/3546 
- https://github.com/snowflakedb/terraform-provider-snowflake/pull/3547
- https://github.com/snowflakedb/terraform-provider-snowflake/pull/3548