fix: Handle drop operation safely (#3570)
## Changes
- Add a common drop function (with tests) that accounts for missing
high-hierarchy objects
  - I added the implicit `.WithIfExists(true)` option
- Add DropSafely method that uses a common drop function
- Test errors that may occur for drop operations with missing
high-hierarchy objects
- Add new error to distinct SDK error that is skippable from underlying
Snowflake ones (check ideas section)
- Add test implementation for tables and streamlits with tests
- Add common delete function for resoruces (wanted to keep it as short
and simple as possible, but still ended up with two functions that need
to be passed; the only solution that I thought of to shorten the
function further is replacing parameters with resource name and map to
correct arguments (parsing and Drop function) inside the common delete
function basing on resource name. It could be less safe approach,
because of the missing mappings, but maybe a cleaner and better one).
- For more complex drop operations (like for accounts), I would keep the
dedicated drop function

## Next steps
- Adjust templates to include DropSafely method
- Move common "safe" drop and show functions to one file
- Use in all resources
- Go through object renaming tests and see how the changes with drop and
read affect those

## Ideas
- Use ErrSkippable in the safe show function (like in the current safe
drop implementation) to distinct ErrObjectNotFound (error form the
collections.Find) from the actual skippable operation.