fix: Network policy test fixes (#3574)
## Changes
- Adjust tests to account for new network policy
- Change default value for user network policy parameter to
RESTRICTED_ACCESS and change it's default level to ACCOUNT
- Adjust CheckUserPasswordPolicyAttachmentDestroy logic to only check
for password policies (not all policies as it was previously doing)
- Remove TestAcc_UserPasswordPolicyAttachment_gh3005 as it was referring
to the old provider code (v0.87.0), which had a bug and was mapping
POLICY_STATUS as a non-null field. It was fixed later on and it will be
always failing because of the network policy on the account that has
always null status (I'm guessing status for this policy doesn't matter).

## Next steps
- Adjust parameter asserts, particularly HasAllDefaults function to
check for default values on the level by checking with locally defined
default value (in the object_parameters_def.go) rather than from
Snowflake. Because of the current behavior, the network policy would
always fail, because the Snowflake default is set to an empty string,
not RESTRICTED_ACCESS, as it is in our parameter definitions.
- Create a follow-up ticket and put it into commented out
`HasDefaultParameterValueOnLevel(sdk.UserParameterNetworkPolicy,
sdk.ParameterTypeAccount)`